Refactor: Refactoring functionality to delete a user

+++ src/main/java/com/api/scholarships/services/implementation/UserServiceImp.java
@@ -85,6 +85,9 @@ public class UserServiceImp implements UserService {
   public void delete(Long id) {
     // TODO:verify if an user with a company it's possible delete
     User userFound = getById(id);
+    if(userFound.getCompany()!= null){
+      throw new BadRequestException(Messages.MESSAGE_CANNOT_DELETE_USER);
+    }
     userRepository.delete(userFound);
   }
